Use cached Path.path where possible for significant speed gain
The trade-off is that we need to cache the path as an attribute of the Path
instance itself; otherwise, if we were to use that instance as a key into some
other cache, that would result in infinite recursion for Path.__hash__ (which
we also want to benefit from the caching).

The memory overhead of this is low, because the path string gets interned, and
will be shared with any childrens' parent attribute.

Therefore, the only significant drawback of the path atrtibute would seem to be
the redundancy of Path.__repr__ output.